+++ crates/store/fuzz/src/store.rs
@@ -15,8 +15,8 @@
 use std::collections::HashMap;
 use std::convert::TryInto;
 
-use rand_core::{RngCore, SeedableRng};
 use rand_pcg::Pcg32;
+use rand_pcg::rand_core::{RngCore, SeedableRng};
 use wasefire_store::{
     BufferOptions, BufferStorage, Store, StoreDriver, StoreDriverOff, StoreDriverOn,
     StoreInterruption, StoreInvariant, StoreOperation, StoreUpdate,