+++ crates/prelude/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 0.8.1-git
+## 0.8.1
 
 ### Minor
 
@@ -221,4 +221,4 @@
 
 ## 0.1.0
 
-<!-- Increment to skip CHANGELOG.md test: 8 -->
+<!-- Increment to skip CHANGELOG.md test: 0 -->